+++ src/main/java/subway/application/LineService.java
@@ -3,8 +3,8 @@ package subway.application;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.application.dto.LineCreateDto;
-import subway.application.dto.LineUpdateDto;
+import subway.application.dto.line.LineCreateDto;
+import subway.application.dto.line.LineUpdateDto;
 import subway.domain.Line;
 import subway.domain.repository.LineRepository;
 import subway.exception.IllegalLineException;
