+++ src/main/java/subway/application/PathService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.application.dto.PathFindDto;
-import subway.application.dto.ShortestPathDto;
+import subway.application.dto.path.PathFindDto;
+import subway.application.dto.path.ShortestPathDto;
 import subway.domain.FareCalculator;
 import subway.domain.Line;
 import subway.domain.Path;
