+++ src/main/java/subway/application/SectionService.java
@@ -6,15 +6,15 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.application.dto.SectionCreateDto;
-import subway.application.dto.SectionDeleteDto;
+import subway.application.dto.section.SectionCreateDto;
+import subway.application.dto.section.SectionDeleteDto;
 import subway.domain.Line;
 import subway.domain.Section;
 import subway.domain.Station;
 import subway.domain.repository.LineRepository;
 import subway.domain.repository.SectionRepository;
 import subway.domain.repository.StationRepository;
-import subway.dto.section.SectionResponse;
+import subway.ui.dto.section.SectionResponse;
 import subway.exception.IllegalDistanceException;
 import subway.exception.IllegalSectionException;
 import subway.exception.StationNotFoundException;
