+++ src/main/java/subway/application/StationService.java
@@ -3,8 +3,8 @@ package subway.application;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.application.dto.StationCreateDto;
-import subway.application.dto.StationUpdateDto;
+import subway.application.dto.station.StationCreateDto;
+import subway.application.dto.station.StationUpdateDto;
 import subway.domain.Station;
 import subway.domain.repository.StationRepository;
 import subway.exception.IllegalStationException;
