+++ src/main/java/subway/application/dto/line/LineCreateDto.java
@@ -1,18 +1,12 @@
-package subway.application.dto;
+package subway.application.dto.line;
 
-public class LineUpdateDto {
-    private Long id;
+public class LineCreateDto {
     private String name;
     private String color;
 
-    public LineUpdateDto(Long id, String name, String color) {
-        this.id = id;
+    public LineCreateDto(String name, String color) {
         this.name = name;
         this.color = color;
-    }
-
-    public Long getId() {
-        return id;
     }
 
     public String getName() {
