+++ src/main/java/subway/application/dto/line/LineUpdateDto.java
@@ -1,12 +1,18 @@
-package subway.application.dto;
+package subway.application.dto.line;
 
-public class LineCreateDto {
+public class LineUpdateDto {
+    private Long id;
     private String name;
     private String color;
 
-    public LineCreateDto(String name, String color) {
+    public LineUpdateDto(Long id, String name, String color) {
+        this.id = id;
         this.name = name;
         this.color = color;
+    }
+
+    public Long getId() {
+        return id;
     }
 
     public String getName() {
