+++ src/main/java/subway/application/dto/path/ShortestPathDto.java
@@ -1,4 +1,4 @@
-package subway.application.dto;
+package subway.application.dto.path;
 
 import java.util.List;
 import subway.domain.Path;
