+++ src/main/java/subway/application/dto/section/SectionCreateDto.java
@@ -1,4 +1,4 @@
-package subway.application.dto;
+package subway.application.dto.section;
 
 public class SectionCreateDto {
     private Long lineId;
