+++ src/main/java/subway/ui/dto/line/LineCreateRequest.java
@@ -1,27 +1,26 @@
-package subway.dto.line;
+package subway.ui.dto.line;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import javax.validation.constraints.NotBlank;
-import subway.application.dto.LineUpdateDto;
+import subway.application.dto.line.LineCreateDto;
 
-public class LineUpdateRequest {
+public class LineCreateRequest {
     @NotBlank(message = "노선의 이름은 비어있을 수 없습니다.")
     private String name;
 
     @NotBlank(message = "노선의 색은 비어있을 수 없습니다.")
     private String color;
 
-    public LineUpdateRequest() {
-    }
+    public LineCreateRequest() {}
 
     @JsonCreator
-    public LineUpdateRequest(String name, String color) {
+    public LineCreateRequest(String name, String color) {
         this.name = name;
         this.color = color;
     }
 
-    public LineUpdateDto toLineUpdateDto(Long id) {
-        return new LineUpdateDto(id, name, color);
+    public LineCreateDto toLineCreateDto() {
+        return new LineCreateDto(name, color);
     }
 
     public String getName() {
