+++ src/main/java/subway/ui/dto/section/SectionResponse.java
@@ -1,4 +1,4 @@
-package subway.dto.section;
+package subway.ui.dto.section;
 
 import subway.domain.Section;
