+++ src/test/java/subway/integration/StationControllerIntegrationTest.java
@@ -14,7 +14,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
-import subway.dto.station.StationCreateRequest;
+import subway.ui.dto.station.StationCreateRequest;
 
 @SpringBootTest
 @Transactional
