+++ src/test/java/subway/ui/LineControllerTest.java
@@ -14,9 +14,9 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import subway.application.LineService;
-import subway.application.dto.LineCreateDto;
+import subway.application.dto.line.LineCreateDto;
 import subway.domain.Line;
-import subway.dto.line.LineCreateRequest;
+import subway.ui.dto.line.LineCreateRequest;
 
 @WebMvcTest(LineController.class)
 class LineControllerTest {
